pyproject.toml
[project]
name = "RL-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "CooperBigFoot", email = "nlazaro@student.ethz.ch" }]
requires-python = ">=3.10"
dependencies = [
    "gymnasium>=1.2.0",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.3.0",
    "pyyaml>=6.0.2",
    "tensorboard>=2.19.0",
    "tensordict>=0.8.3",
    "torch>=2.7.1",
    "torchrl>=0.8.1",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = ["pytest>=8.3.5", "ruff>=0.11.8", "ty>=0.0.1a12"]

[tool.setuptools.packages.find]
where = ["src"]

# RUFF Configuration
[tool.ruff]
line-length = 120
target-version = "py310"
fix = false
exclude = ["notebooks/"]

[tool.ruff.lint]
# Essential rules: errors, warnings, naming, imports, modern Python, bugs, comprehensions, simplifications
select = ["E", "F", "W", "N", "I", "UP", "B", "C4", "SIM"]
ignore = [
    "E501", # Line too long (let formatter handle this)
    "N803", # Argument name should be lowercase - ML convention allows X, Y
    "N806", # Variable in function should be lowercase - ML convention allows X, Y
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# TY Configuration  
[tool.ty]
# Ty doesn't support exclude in pyproject.toml yet
# Files are excluded via .gitignore or command line args

[tool.ty.rules]
# Example: Set some rules to warn instead of error for gradual adoption
possibly-unresolved-reference = "warn"

# Pytest Configuration
[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",                 # Show all test results
    "--ignore=notebooks/",
]
